ChatCommand moved to Commands dir
