+++ Explorer/Assets/DCL/Minimap/MinimapController.cs
@@ -5,7 +5,7 @@ using Arch.SystemGroups.DefaultSystemGroups;
 using Cysharp.Threading.Tasks;
 using DCL.Character.CharacterMotion.Components;
 using DCL.Character.Components;
-using DCL.Chat;
+using DCL.Chat.MessageBus;
 using DCL.Diagnostics;
 using DCL.ExplorePanel;
 using DCL.MapRenderer;
@@ -45,10 +45,8 @@ namespace DCL.Minimap
         private MapRendererTrackPlayerPosition mapRendererTrackPlayerPosition;
         private IMapCameraController mapCameraController;
         private Vector2Int previousParcelPosition;
-        private SideMenuController sideMenuController;
         private readonly IRealmNavigator realmNavigator;
         private readonly IScenesCache scenesCache;
-
 
         public IReadOnlyDictionary<MapLayer, IMapLayerParameter> LayersParameters { get; } = new Dictionary<MapLayer, IMapLayerParameter>
             { { MapLayer.PlayerMarker, new PlayerMarkerParameter { BackgroundIsActive = false } } };
@@ -92,7 +90,7 @@ namespace DCL.Minimap
             viewInstance.goToGenesisCityButton.onClick.AddListener(() => chatMessagesBus.Send("/goto 0,0"));
             viewInstance.SideMenuCanvasGroup.alpha = 0;
             viewInstance.SideMenuCanvasGroup.gameObject.SetActive(false);
-            sideMenuController = new SideMenuController(viewInstance.sideMenuView);
+            new SideMenuController(viewInstance.sideMenuView);
             SetWorldMode(realmData.ScenesAreFixed);
             realmNavigator.RealmChanged += OnRealmChanged;
         }
@@ -115,10 +113,7 @@ namespace DCL.Minimap
 
         private void OpenSideMenu()
         {
-            if (viewInstance.SideMenuCanvasGroup.gameObject.activeInHierarchy)
-            {
-                viewInstance.SideMenuCanvasGroup.DOFade(0, ANIMATION_TIME).SetEase(Ease.InOutQuad).OnComplete(() => viewInstance.SideMenuCanvasGroup.gameObject.gameObject.SetActive(false));
-            }
+            if (viewInstance.SideMenuCanvasGroup.gameObject.activeInHierarchy) { viewInstance.SideMenuCanvasGroup.DOFade(0, ANIMATION_TIME).SetEase(Ease.InOutQuad).OnComplete(() => viewInstance.SideMenuCanvasGroup.gameObject.gameObject.SetActive(false)); }
             else
             {
                 viewInstance.SideMenuCanvasGroup.gameObject.gameObject.SetActive(true);
@@ -183,7 +178,7 @@ namespace DCL.Minimap
             bool isNotEmptyParcel = scenesCache.Contains(playerParcelPosition);
             bool isSdk7Scene = scenesCache.TryGetByParcel(playerParcelPosition, out _);
             viewInstance.sdk6Label.gameObject.SetActive(isNotEmptyParcel && !isSdk7Scene);
-            
+
             return;
 
             async UniTaskVoid RetrieveParcelInfoAsync(Vector2Int playerParcelPosition)
@@ -199,7 +194,6 @@ namespace DCL.Minimap
                         PlacesData.PlaceInfo? placeInfo = await placesAPIService.GetPlaceAsync(playerParcelPosition, cts.Token);
                         viewInstance.placeNameText.text = placeInfo?.title ?? "Unknown place";
                     }
-                    
                 }
                 catch (NotAPlaceException notAPlaceException)
                 {
@@ -207,10 +201,7 @@ namespace DCL.Minimap
                     ReportHub.LogWarning(ReportCategory.UNSPECIFIED, $"Not a place requested: {notAPlaceException.Message}");
                 }
                 catch (Exception) { viewInstance.placeNameText.text = "Unknown place"; }
-                finally
-                {
-                    viewInstance.placeCoordinatesText.text = playerParcelPosition.ToString().Replace("(", "").Replace(")", "");
-                }
+                finally { viewInstance.placeCoordinatesText.text = playerParcelPosition.ToString().Replace("(", "").Replace(")", ""); }
             }
         }
