+++ Explorer/Assets/Scripts/Global/Dynamic/ChatCommands/DebugPanelChatCommand.cs
@@ -1,5 +1,5 @@
 ﻿using Cysharp.Threading.Tasks;
-using DCL.Chat;
+using DCL.Chat.Commands;
 using DCL.DebugUtilities;
 using DCL.DebugUtilities.Views;
 using System.Text.RegularExpressions;
